requirements.txt
langchain
langchain-groq
langchain-openai
langchain-chroma
langchain-community
faker
python-dotenv
streamlit
watchdog
pypdf